Deprecate caddy command in favor of ingress command
