+++ cmd/caddy/init.go
@@ -1,16 +1,10 @@
 package caddy
 
 import (
-	"bytes"
 	"fmt"
-	"io"
 	"os"
-	"os/exec"
-	"sync"
-	"time"
 
-	"github.com/bitswan-space/bitswan-workspaces/internal/caddyapi"
-	"github.com/bitswan-space/bitswan-workspaces/internal/dockercompose"
+	"github.com/bitswan-space/bitswan-workspaces/cmd/ingress"
 	"github.com/spf13/cobra"
 )
 
@@ -23,7 +17,9 @@ func newInitCmd() *cobra.Command {
 		Short: "Initializes a Caddy",
 		Args:  cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if err := InitCaddy(domain, verbose); err != nil {
+			// Print deprecation warning to STDERR
+			fmt.Fprintln(os.Stderr, "WARNING: The 'caddy init' command is deprecated and will be removed in a future version. Please use 'ingress init' instead.")
+			if err := ingress.InitIngress(domain, verbose); err != nil {
 				return fmt.Errorf("failed to initialize Caddy: %w", err)
 			}
 			return nil
@@ -38,140 +34,4 @@ func newInitCmd() *cobra.Command {
 	return cmd
 }
 
-func runCommandVerbose(cmd *exec.Cmd, verbose bool) error {
-	var stdoutBuf, stderrBuf bytes.Buffer
 
-	if verbose {
-		// Set up pipes for real-time streaming
-		stdoutPipe, err := cmd.StdoutPipe()
-		if err != nil {
-			return fmt.Errorf("failed to create stdout pipe: %w", err)
-		}
-
-		stderrPipe, err := cmd.StderrPipe()
-		if err != nil {
-			return fmt.Errorf("failed to create stderr pipe: %w", err)
-		}
-
-		// Create multi-writers to both stream and capture output
-		stdoutWriter := io.MultiWriter(os.Stdout, &stdoutBuf)
-		stderrWriter := io.MultiWriter(os.Stderr, &stderrBuf)
-
-		// Start the command
-		if err := cmd.Start(); err != nil {
-			return fmt.Errorf("failed to start command: %w", err)
-		}
-
-		// Copy stdout and stderr in separate goroutines
-		var wg sync.WaitGroup
-		wg.Add(2)
-
-		go func() {
-			defer wg.Done()
-			io.Copy(stdoutWriter, stdoutPipe)
-		}()
-
-		go func() {
-			defer wg.Done()
-			io.Copy(stderrWriter, stderrPipe)
-		}()
-
-		// Wait for all output to be processed
-		wg.Wait()
-
-		// Wait for command to complete
-		err = cmd.Wait()
-		return err
-	} else {
-		// Not verbose, just capture output for potential error reporting
-		cmd.Stdout = &stdoutBuf
-		cmd.Stderr = &stderrBuf
-
-		err := cmd.Run()
-
-		// If command failed, print the captured output
-		if err != nil {
-			if stdoutBuf.Len() > 0 {
-				fmt.Println("Command stdout:")
-				fmt.Println(stdoutBuf.String())
-			}
-
-			if stderrBuf.Len() > 0 {
-				fmt.Println("Command stderr:")
-				fmt.Println(stderrBuf.String())
-			}
-		}
-
-		return err
-	}
-}
-
-func InitCaddy(domain string, verbose bool) error {
-	bitswanConfig := os.Getenv("HOME") + "/.config/bitswan/"
-	caddyConfig := bitswanConfig + "caddy"
-	caddyCertsDir := caddyConfig + "/certs"
-
-	fmt.Println("Setting up Caddy...")
-	if err := os.MkdirAll(caddyConfig, 0755); err != nil {
-		return fmt.Errorf("failed to create Caddy config directory: %w", err)
-	}
-
-	// Create Caddyfile with email and modify admin listener
-	caddyfile := `
-		{
-			email info@bitswan.space
-			admin 0.0.0.0:2019
-		}`
-
-	caddyfilePath := caddyConfig + "/Caddyfile"
-	if err := os.WriteFile(caddyfilePath, []byte(caddyfile), 0755); err != nil {
-		panic(fmt.Errorf("failed to write Caddyfile: %w", err))
-	}
-
-	caddyDockerCompose, err := dockercompose.CreateCaddyDockerComposeFile(caddyConfig, domain)
-	if err != nil {
-		panic(fmt.Errorf("failed to create Caddy docker-compose file: %w", err))
-	}
-
-	caddyDockerComposePath := caddyConfig + "/docker-compose.yml"
-	if err := os.WriteFile(caddyDockerComposePath, []byte(caddyDockerCompose), 0755); err != nil {
-		panic(fmt.Errorf("failed to write Caddy docker-compose file: %w", err))
-	}
-
-	err = os.Chdir(caddyConfig)
-	if err != nil {
-		panic(fmt.Errorf("failed to change directory to Caddy config: %w", err))
-	}
-
-	caddyProjectName := "bitswan-caddy"
-	caddyDockerComposeCom := exec.Command("docker", "compose", "-p", caddyProjectName, "up", "-d")
-
-	// Capture both stdout and stderr
-	var stdout, stderr bytes.Buffer
-	caddyDockerComposeCom.Stdout = &stdout
-	caddyDockerComposeCom.Stderr = &stderr
-
-	// Create certs directory if it doesn't exist
-	if _, err := os.Stat(caddyCertsDir); os.IsNotExist(err) {
-		if err := os.MkdirAll(caddyCertsDir, 0740); err != nil {
-			return fmt.Errorf("failed to create Caddy certs directory: %w", err)
-		}
-	}
-
-	fmt.Println("Starting Caddy...")
-	if err := runCommandVerbose(caddyDockerComposeCom, verbose); err != nil {
-		// Combine stdout and stderr for complete output
-		fullOutput := stdout.String() + stderr.String()
-		return fmt.Errorf("failed to start Caddy:\nError: %v\nOutput:\n%s", err, fullOutput)
-	}
-
-	// wait 5s to make sure Caddy is up
-	time.Sleep(5 * time.Second)
-	err = caddyapi.InitCaddy()
-	if err != nil {
-		panic(fmt.Errorf("failed to init Caddy: %w", err))
-	}
-
-	fmt.Println("Caddy started successfully!")
-	return nil
-}
