+++ cmd/ingress/root.go
@@ -1,0 +1,17 @@
+package ingress
+
+import "github.com/spf13/cobra"
+
+func NewIngressCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "ingress",
+		Short: "Manage ingress",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return cmd.Help()
+		},
+	}
+
+	cmd.AddCommand(newInitCmd())
+
+	return cmd
+}
