+++ cmd/root.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/bitswan-space/bitswan-workspaces/cmd/automation"
 	"github.com/bitswan-space/bitswan-workspaces/cmd/caddy"
+	"github.com/bitswan-space/bitswan-workspaces/cmd/ingress"
 	"github.com/spf13/cobra"
 )
 
@@ -24,7 +25,8 @@ func newRootCmd(version string) *cobra.Command {
 	cmd.AddCommand(newVersionCmd(version)) // version subcommand
 	cmd.AddCommand(newWorkspaceCmd())      // workspace subcommand
 	cmd.AddCommand(newRegisterCmd())       // register subcommand
-	cmd.AddCommand(caddy.NewCaddyCmd())    // caddy subcommand
+	cmd.AddCommand(ingress.NewIngressCmd()) // ingress subcommand
+	cmd.AddCommand(caddy.NewCaddyCmd())    // caddy subcommand (deprecated)
 
 	// Check if the configuration file exists and has an active workspace
 	configPath := filepath.Join(os.Getenv("HOME"), ".config", "bitswan", "config.toml")
